:art: flatten blog structure
