+++ src/pages/blog/a-better-album-ranking-algorithm.mdx
@@ -1,4 +1,4 @@
-import BlogPost from '../../../components/BlogPost'
+import BlogPost from '../../components/BlogPost'
 
 export const meta = {
   title: 'A Better Album Ranking Algorithm',
