+++ src/pages/blog/a-form-fitted-z-shell.mdx
@@ -1,4 +1,4 @@
-import BlogPost from '../../../components/BlogPost'
+import BlogPost from '../../components/BlogPost'
 
 export const meta = {
   title: 'A Form-Fitted Z Shell',
