+++ src/pages/blog/components-mdx-secret-weapon.mdx
@@ -1,6 +1,6 @@
-import BlogPost from '../../../components/BlogPost'
-import Link from '../../../components/Link'
-import ExampleCounter from './ExampleCounter'
+import BlogPost from '../../components/BlogPost'
+import Link from '../../components/Link'
+import ExampleCounter from '../../components/MDXExampleCounter'
 
 export const meta = {
   title: "Components are MDX's Secret Weapon",
