+++ src/pages/blog/customizing-tachyons.mdx
@@ -1,4 +1,4 @@
-import BlogPost from '../../../components/BlogPost'
+import BlogPost from '../../components/BlogPost'
 
 export const meta = {
   title: ' Customizing Tachyons',
