+++ src/pages/blog/fetch-with-timeout.mdx
@@ -1,4 +1,4 @@
-import BlogPost from '../../../components/BlogPost'
+import BlogPost from '../../components/BlogPost'
 
 export const meta = {
   title: 'Request Timeouts With the Fetch API',
