+++ src/pages/blog/functional-first-accordions.mdx
@@ -1,6 +1,6 @@
 import { Text, Container, Link } from 'theme-ui'
-import Update from '../../../components/Update'
-import BlogPost from '../../../components/BlogPost'
+import Update from '../../components/Update'
+import BlogPost from '../../components/BlogPost'
 
 export const meta = {
   title: 'Functional-First Accordions',
