+++ src/pages/blog/no-red-squigglies-for-semicolons.mdx
@@ -1,4 +1,4 @@
-import BlogPost from '../../../components/BlogPost'
+import BlogPost from '../../components/BlogPost'
 
 export const meta = {
   title: 'No Red Squigglies For Semicolons',
